refactor event mapper
