+++ icu_pipeline/mapper/source/mimic/labevents.py
@@ -1,47 +1,10 @@
 from abc import ABCMeta
 
-import pandas as pd
-from pandera.typing import DataFrame
-
-from icu_pipeline.mapper.source import AbstractDatabaseSourceMapper
-from icu_pipeline.mapper.schema.ohdsi import AbstractOHDSISinkSchema
-from icu_pipeline.mapper.schema.fhir import (
-    Identifier,
-    Subject,
-    ValueQuantity,
-    FHIRObservation,
-)
+from icu_pipeline.mapper.source.mimic import AbstractMimicEventsMapper
 
 
-class AbstractMimicLabEventsMapper(
-    AbstractDatabaseSourceMapper[FHIRObservation, AbstractOHDSISinkSchema],
-    metaclass=ABCMeta,
-):
+class AbstractMimicLabEventsMapper(AbstractMimicEventsMapper, metaclass=ABCMeta):
     SQL_QUERY = "SELECT * FROM mimiciv_hosp.labevents WHERE itemid = any(%(values)s);"
-
-    def _to_fihr(self, df: DataFrame) -> DataFrame[FHIRObservation]:
-        observation_df = pd.DataFrame()
-
-        observation_df[FHIRObservation.subject] = df["subject_id"].map(
-            lambda id: Subject(reference=str(id))
-        )
-        observation_df[FHIRObservation.effective_date_time] = pd.to_datetime(
-            df["charttime"], utc=True
-        )
-        observation_df[FHIRObservation.value_quantity] = df.apply(
-            lambda _df: ValueQuantity(
-                value=float(_df["valuenum"]), unit=_df["valueuom"]
-            ),
-            axis=1,
-        )
-        observation_df[FHIRObservation.identifier] = [
-            Identifier(value=self._snomed_id)
-        ] * len(df)
-
-        return observation_df.pipe(DataFrame[FHIRObservation])
-
-    def _to_ohdsi(self, df: DataFrame) -> DataFrame[AbstractOHDSISinkSchema]:
-        raise NotImplementedError
 
 
 class MimicSerumCreatinineMapper(AbstractMimicLabEventsMapper):
